+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features added:
 * Cover MCL in mapping profile details sections with tests (UIDATIMP-469)
 * Implement <ProhibitionIcon> component (UIDATIMP-477)
+* Distinguish not mapped and unmappable fields in Field mapping profile View screen (UIDATIMP-471)
 
 ### Bugs fixed:
 * When returned to search results screen after profile save, wrong profile details show (UIDATIMP-424)